LayoutTests/media/media-source/media-webm-opus-codecdelay.html

<!DOCTYPE html>
<html>
<head>
    <title>media-webm-opus-codecdelay</title>
    <script src="media-source-loader.js"></script>
    <script src="../video-test.js"></script>
    <script>
    var loader;
    var source;
    var sourceBuffer;
    var random;

    function loaderPromise(loader) {
        return new Promise((resolve, reject) => {
            loader.onload = resolve;
            loader.onerror = reject;
        });
    }

    window.addEventListener('load', async event => {
        try {
            findMediaElement();
            loader = new MediaSourceLoader('content/test-opus-codecdelay-manifest.json');
            await loaderPromise(loader);

            source = new MediaSource();
            run('video.src = URL.createObjectURL(source)');
            await waitFor(source, 'sourceopen');
            waitFor(video, 'error').then(failTest);

            consoleWrite('Append the init segment');
            run('sourceBuffer = source.addSourceBuffer(loader.type())');
            run('init = loader.initSegment()');
            run('sourceBuffer.appendBuffer(init)');
            await waitFor(sourceBuffer, 'update');

            consoleWrite('Append a partial media segment');
            run('sourceBuffer.appendBuffer(loader.mediaSegment(0).slice(0, loader.mediaSegmentSize(0)/2))');
            await waitFor(sourceBuffer, 'update');
            testExpected('sourceBuffer.buffered.length', '1');
            testExpected('sourceBuffer.buffered.end(0)', '3.9545');

            endTest();
        } catch (e) {
            failTest(`Caught exception: "${e}"`);
        }
    });
    </script>
</head>
<body>
    <video controls></video>
</body>
</html>
